Update dnsjava, use desugaring with nio, refactor DNS resolving
