+++ app/src/main/kotlin/at/bitfire/davdroid/network/Android10Resolver.kt
@@ -11,21 +11,21 @@ import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.asExecutor
 import kotlinx.coroutines.runBlocking
+import org.xbill.DNS.EDNSOption
 import org.xbill.DNS.Message
 import org.xbill.DNS.Resolver
-import org.xbill.DNS.ResolverListener
 import org.xbill.DNS.TSIG
 import java.io.IOException
+import java.time.Duration
 
 /**
  * dnsjava Resolver that uses Android's [DnsResolver] API, which is available since Android 10.
  */
 @RequiresApi(Build.VERSION_CODES.Q)
-object Android10Resolver: Resolver {
+class Android10Resolver : Resolver {
 
     private val executor = Dispatchers.IO.asExecutor()
     private val resolver = DnsResolver.getInstance()
-
 
     override fun send(query: Message): Message = runBlocking {
         val future = CompletableDeferred<Message>()
@@ -44,10 +44,6 @@ object Android10Resolver: Resolver {
         future.await()
     }
 
-    override fun sendAsync(query: Message, listener: ResolverListener) =
-        // currently not used by dnsjava, so no need to implement it
-        throw NotImplementedError()
-
 
     override fun setPort(port: Int) {
         // not applicable
@@ -61,11 +57,7 @@ object Android10Resolver: Resolver {
         // not applicable
     }
 
-    override fun setEDNS(level: Int) {
-        // not applicable
-    }
-
-    override fun setEDNS(level: Int, payloadSize: Int, flags: Int, options: MutableList<Any?>?) {
+    override fun setEDNS(version: Int, payloadSize: Int, flags: Int, options: MutableList<EDNSOption>?) {
         // not applicable
     }
 
@@ -73,11 +65,7 @@ object Android10Resolver: Resolver {
         // not applicable
     }
 
-    override fun setTimeout(secs: Int, msecs: Int) {
-        // not applicable
-    }
-
-    override fun setTimeout(secs: Int) {
+    override fun setTimeout(timeout: Duration?) {
         // not applicable
     }
