+++ app/src/main/kotlin/at/bitfire/davdroid/servicedetection/DavResourceFinder.kt
@@ -30,11 +30,14 @@ import at.bitfire.dav4jvm.property.webdav.ResourceType
 import at.bitfire.davdroid.db.Collection
 import at.bitfire.davdroid.db.Credentials
 import at.bitfire.davdroid.log.StringHandler
+import at.bitfire.davdroid.network.DnsRecordResolver
 import at.bitfire.davdroid.network.HttpClient
-import at.bitfire.davdroid.util.DavUtils
+import dagger.assisted.Assisted
+import dagger.assisted.AssistedFactory
+import dagger.assisted.AssistedInject
+import dagger.hilt.android.qualifiers.ApplicationContext
 import okhttp3.HttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
-import org.xbill.DNS.Lookup
 import org.xbill.DNS.Type
 import java.io.InterruptedIOException
 import java.net.SocketTimeoutException
@@ -56,11 +59,17 @@ import java.util.logging.Logger
  * @param baseURI        user-given base URI (either mailto: URI or http(s):// URL)
  * @param credentials    optional login credentials (username/password, client certificate, OAuth state)
  */
-class DavResourceFinder(
-    val context: Context,
-    private val baseURI: URI,
-    private val credentials: Credentials? = null
+class DavResourceFinder @AssistedInject constructor(
+    @Assisted private val baseURI: URI,
+    @Assisted private val credentials: Credentials? = null,
+    @ApplicationContext val context: Context,
+    private val dnsRecordResolver: DnsRecordResolver
 ): AutoCloseable {
+
+    @AssistedFactory
+    interface Factory {
+        fun resourceFinder(baseURI: URI, credentials: Credentials?): DavResourceFinder
+    }
 
     enum class Service(val wellKnownName: String) {
         CALDAV("caldav"),
@@ -385,9 +394,8 @@ class DavResourceFinder(
         val query = "_${service.wellKnownName}s._tcp.$domain"
         log.fine("Looking up SRV records for $query")
 
-        val srvLookup = Lookup(query, Type.SRV)
-        DavUtils.prepareLookup(context, srvLookup)
-        val srv = DavUtils.selectSRVRecord(srvLookup.run().orEmpty())
+        val srvRecords = dnsRecordResolver.resolve(query, Type.SRV)
+        val srv = dnsRecordResolver.bestSRVRecord(srvRecords)
 
         if (srv != null) {
             // choose SRV record to use (query may return multiple SRV records)
@@ -404,9 +412,8 @@ class DavResourceFinder(
         }
 
         // look for TXT record too (for initial context path)
-        val txtLookup = Lookup(query, Type.TXT)
-        DavUtils.prepareLookup(context, txtLookup)
-        paths.addAll(DavUtils.pathsFromTXTRecords(txtLookup.run()))
+        val txtRecords = dnsRecordResolver.resolve(query, Type.TXT)
+        paths.addAll(dnsRecordResolver.pathsFromTXTRecords(txtRecords))
 
         // in case there's a TXT record, but it's wrong, try well-known
         paths.add("/.well-known/" + service.wellKnownName)
