+++ gradle/libs.versions.toml
@@ -24,8 +24,7 @@ bitfire-ical4android = "83cda23ceb"
 bitfire-vcard4android = "03a37a8284"
 compose-accompanist = "0.34.0"
 compose-bom = "2024.06.00"
-# don't update to dnsjava 3.x until API level 26 (Android 8) is the minimum API [https://github.com/bitfireAT/davx5/issues/453]
-dnsjava = "2.1.9"
+dnsjava = "3.5.3"
 glance = "1.1.0"
 guava = "33.2.1-android"
 hilt = "2.51.1"
@@ -44,7 +43,7 @@ room = "2.6.1"
 unifiedpush = "2.4.0"
 
 [libraries]
-android-desugaring = { module = "com.android.tools:desugar_jdk_libs", version.ref = "android-desugaring" }
+android-desugaring = { module = "com.android.tools:desugar_jdk_libs_nio", version.ref = "android-desugaring" }
 androidx-activityCompose = { module = "androidx.activity:activity-compose", version.ref = "androidx-activityCompose" }
 androidx-appcompat = { module = "androidx.appcompat:appcompat", version.ref = "androidx-appcompat" }
 androidx-arch-core-testing = { module = "androidx.arch.core:core-testing", version.ref = "androidx-arch" }
